+++ source/src/dict/Ostap.hh
@@ -40,6 +40,7 @@
 #include "Ostap/DataFrameUtils.h"
 #include "Ostap/Differences.h"
 #include "Ostap/Digit.h"
+#include "Ostap/EffEntity.h"
 #include "Ostap/ECDF.h"
 #include "Ostap/EPDF.h"
 #include "Ostap/EigenSystem.h"